ml_models/flare_predictor/requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.21.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.26.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
scikit-learn>=0.24.2
nltk>=3.6.3
pytest>=6.2.5
logging>=0.5.1.2

# NLP and text processing
spacy>=3.5.0

# Database
SQLAlchemy>=1.4.0

# HTTP and API
aiohttp>=3.8.0

# Data generation and anonymization
Faker>=8.0.0
cryptography>=3.4.0

# Configuration and environment
tqdm>=4.62.0
joblib>=1.1.0

# Development and testing
black>=21.5b2
flake8>=3.9.0
mypy>=0.910 